Make a cleanup
